fix: add type casting

+++ animation/Blending.cpp
@@ -6,7 +6,7 @@ namespace anim {
 		if (parentBone == boneToCheck) { return true; }
 		int parent = pose.ParentIndexOf(boneToCheck);
 		while (parent >= 0) {
-			if (parent == parentBone) { return true; }
+			if (parent == (int) parentBone) { return true; }
 			parent = pose.ParentIndexOf(parent);
 		}
 		return false;
@@ -16,7 +16,7 @@ namespace anim {
 		unsigned int numbBones = poseOut.Size();
 		for (unsigned int bone = 0; bone < numbBones; bone++) {
 			// negative root bone is assume to mean the caller wants blending applied to every bone
-			if (rootBone >= 0 && !IsBoneChildOf(poseOut, rootBone, bone)) { continue; }
+			if (rootBone >= 0 && !IsBoneChildOf(poseOut, (unsigned int) rootBone, bone)) { continue; }
 			// blend bones in local pose space
 			poseOut.SetLocalTransform(bone,
 				transforms::mix(a.GetLocalTransform(bone), b.GetLocalTransform(bone), t)
